+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 **Version 2.1**
 
-- Changed implementation of MATCH metafile context function.
+- Changed implementation of MATCH and INLIST metafile context functions.
 - Fixed behavior of AT metafile context function.
 - Updated Groovy library from version 2.5.3 (indy) to version 2.5.4 (indy).
 - Re-added a way to disable pre-compiled edits lookup, it was removed by mistake in the previous release.